Moved Huffman decode into test code
