+++ src/test/Makefile
@@ -189,6 +189,10 @@ $(OBJECT_DIR)/common/huffman.o : $(USER_DIR)/common/huffman.c $(USER_DIR)/common
 	@mkdir -p $(dir $@)
 	$(CC) $(C_FLAGS) $(TEST_CFLAGS) -c $(USER_DIR)/common/huffman.c -o $@
 
+$(OBJECT_DIR)/common/huffman_table.o : $(USER_DIR)/common/huffman_table.c $(USER_DIR)/common/huffman.h $(GTEST_HEADERS)
+	@mkdir -p $(dir $@)
+	$(CC) $(C_FLAGS) $(TEST_CFLAGS) -c $(USER_DIR)/common/huffman_table.c -o $@
+
 $(OBJECT_DIR)/huffman_unittest.o : \
 	$(TEST_DIR)/huffman_unittest.cc \
 	$(USER_DIR)/common/huffman.h \
@@ -199,6 +203,7 @@ $(OBJECT_DIR)/huffman_unittest.o : \
 
 $(OBJECT_DIR)/huffman_unittest : \
 	$(OBJECT_DIR)/common/huffman.o \
+	$(OBJECT_DIR)/common/huffman_table.o \
 	$(OBJECT_DIR)/huffman_unittest.o \
 	$(OBJECT_DIR)/gtest_main.a
